+++ R/extended.R
@@ -19,40 +19,30 @@
 # Visit our website for more info: https://msberends.gitlab.io/AMR.    #
 # ==================================================================== #
 
-context("data.R")
+#' Extended functions
+#'
+#' These functions are extensions of functions in other packages.
+#' @inheritSection AMR Read more on our website!
+#' @export
+#' @keywords internal
+#' @name extended-functions
+#' @rdname extended-functions
+#' @exportMethod scale_type.mo
+#' @export
+scale_type.mo <- function(x) {
+  # fix for:
+  # "Don't know how to automatically pick scale for object of type mo. Defaulting to continuous."
+  # "Error: Discrete value supplied to continuous scale"
+  "discrete"
+}
 
-test_that("data sets are valid", {
-  # IDs should always be unique
-  expect_identical(nrow(antibiotics), length(unique(antibiotics$ab)))
-  expect_identical(class(antibiotics$ab), "ab")
-  expect_identical(nrow(microorganisms), length(unique(microorganisms$mo)))
-  expect_identical(class(microorganisms$mo), "mo")
+#' @rdname extended-functions
+#' @exportMethod scale_type.ab
+#' @export
+scale_type.ab <- function(x) {
+  # fix for:
+  # "Don't know how to automatically pick scale for object of type mo. Defaulting to continuous."
+  # "Error: Discrete value supplied to continuous scale"
+  "discrete"
+}
 
-  # check cross table reference
-  expect_true(all(microorganisms.codes$mo %in% microorganisms$mo))
-
-  # antibiotic names must always be coercible to their original AB code
-  expect_identical(antibiotics$ab, as.ab(antibiotics$name))
-
-  # there should be no diacritics (i.e. non ASCII) characters in the datasets (CRAN policy)
-  datasets <- data(package = "AMR", envir = asNamespace("AMR"))$results[, "Item"]
-  for (i in 1:length(datasets)) {
-    dataset <- get(datasets[i], envir = asNamespace("AMR"))
-    expect_identical(dataset_UTF8_to_ASCII(dataset), dataset)
-  }
-})
-
-test_that("creation of data sets is valid", {
-  DT <- make_DT()
-  expect_lt(nrow(DT[prevalence == 1]), nrow(DT[prevalence == 2]))
-  expect_lt(nrow(DT[prevalence == 2]), nrow(DT[prevalence == 3]))
-  old <- make_trans_tbl()
-  expect_gt(length(old), 0)
-})
-
-test_that("CoL version info works", {
- expect_identical(class(catalogue_of_life_version()),
-                  c("catalogue_of_life_version", "list"))
-
-  expect_output(print(catalogue_of_life_version()))
-})