+++ R/mo_property.R
@@ -69,10 +69,11 @@
 #' mo_shortname("E. coli")       # "E. coli"
 #'
 #' ## other properties
-#' mo_gramstain("E. coli")       # "Gram negative"
-#' mo_type("E. coli")            # "Bacteria" (equal to kingdom)
+#' mo_gramstain("E. coli")       # "Gram-negative"
+#' mo_type("E. coli")            # "Bacteria" (equal to kingdom, but may be translated)
 #' mo_rank("E. coli")            # "species"
 #' mo_url("E. coli")             # get the direct url to the online database entry
+#' mo_synonyms("E. coli")        # get previously accepted taxonomic names
 #'
 #' ## scientific reference
 #' mo_ref("E. coli")             # "Castellani et al., 1919"
@@ -84,7 +85,7 @@
 #' mo_genus("MRSA")              # "Staphylococcus"
 #' mo_species("MRSA")            # "aureus"
 #' mo_shortname("MRSA")          # "S. aureus"
-#' mo_gramstain("MRSA")          # "Gram positive"
+#' mo_gramstain("MRSA")          # "Gram-positive"
 #'
 #' mo_genus("VISA")              # "Staphylococcus"
 #' mo_species("VISA")            # "aureus"
@@ -133,15 +134,15 @@
 #'
 #' # get a list with the complete taxonomy (from kingdom to subspecies)
 #' mo_taxonomy("E. coli")
+#' # get a list with the taxonomy, the authors and the URL to the online database
+#' mo_info("E. coli")
 mo_name <- function(x, language = get_locale(), ...) {
-  mo_fullname(x = x, language = language, ... = ...)
+  translate_AMR(mo_validate(x = x, property = "fullname", ...), language = language, only_unknown = FALSE)
 }
 
 #' @rdname mo_property
 #' @export
-mo_fullname <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "fullname", ...), language = language)
-}
+mo_fullname <- mo_name
 
 #' @rdname mo_property
 #' @importFrom dplyr %>% mutate pull
@@ -184,69 +185,61 @@ mo_shortname <- function(x, language = get_locale(), ...) {
   res1[res1 != res2] <- res2_fullname
   result <- as.character(res1)
 
-  t(result, language = language)
+  translate_AMR(result, language = language, only_unknown = FALSE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_subspecies <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "subspecies", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "subspecies", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_species <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "species", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "species", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_genus <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "genus", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "genus", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_family <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "family", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "family", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_order <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "order", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "order", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_class <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "class", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "class", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_phylum <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "phylum", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "phylum", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_kingdom <- function(x, language = get_locale(), ...) {
-  if (all(x %in% AMR::microorganisms$kingdom)) {
-    return(x)
-  }
-  x <- as.mo(x, ...)
-  kngdm <- mo_validate(x = x, property = "kingdom", ...)
-  if (language != "en") {
-    kngdm[x == "UNKNOWN"] <- t(kngdm[x == "UNKNOWN"], language = language)
-  }
-  kngdm
+  translate_AMR(mo_validate(x = x, property = "kingdom", ...), language = language, only_unknown = TRUE)
 }
 
 #' @rdname mo_property
 #' @export
 mo_type <- function(x, language = get_locale(), ...) {
-  t(mo_validate(x = x, property = "kingdom", ...), language = language)
+  translate_AMR(mo_validate(x = x, property = "kingdom", ...), language = language, only_unknown = FALSE)
 }
 
 #' @rdname mo_property
@@ -254,16 +247,26 @@ mo_type <- function(x, language = get_locale(), ...) {
 mo_gramstain <- function(x, language = get_locale(), ...) {
   x.mo <- as.mo(x, ...)
   x.phylum <- mo_phylum(x.mo, language = "en")
+  # DETERMINE GRAM STAIN FOR BACTERIA
+  # Source: https://itis.gov/servlet/SingleRpt/SingleRpt?search_topic=TSN&search_value=956097
+  # It says this:
+  # Kingdom Bacteria (Cavalier-Smith, 2002)
+  #    Subkingdom Posibacteria (Cavalier-Smith, 2002)
+  #   Direct Children:
+  #       Phylum  Actinobacteria (Cavalier-Smith, 2002)
+  #       Phylum  Chloroflexi (Garrity and Holt, 2002)
+  #       Phylum  Firmicutes (corrig. Gibbons and Murray, 1978)
+  #       Phylum  Tenericutes (Murray, 1984)
+  x <- NA_character_
+  # make all bacteria Gram negative
+  x[mo_kingdom(x.mo, language = "en") == "Bacteria"] <- "Gram-negative"
+  # overwrite these phyla with Gram positive
   x[x.phylum %in% c("Actinobacteria",
                     "Chloroflexi",
                     "Firmicutes",
-                    "Tenericutes")] <- "Gram positive"
-  x[x != "Gram positive"] <- "Gram negative"
-  x[mo_kingdom(x.mo, language = "en") != "Bacteria"] <- NA_character_
-  x[x.mo == "B_GRAMP"] <- "Gram positive"
-  x[x.mo == "B_GRAMN"] <- "Gram negative"
-
-  t(x, language = language)
+                    "Tenericutes")
+    | x.mo == "B_GRAMP"] <- "Gram-positive"
+  translate_AMR(x, language = language, only_unknown = FALSE)
 }
 
 #' @rdname mo_property
@@ -276,7 +279,7 @@ mo_ref <- function(x, ...) {
 #' @export
 mo_authors <- function(x, ...) {
   x <- mo_validate(x = x, property = "ref", ...)
-  # remove last 4 digits and presumably the comma and space that preceeds them
+  # remove last 4 digits and presumably the comma and space that preceed them
   x[!is.na(x)] <- gsub(",? ?[0-9]{4}", "", x[!is.na(x)])
   suppressWarnings(x)
 }
@@ -300,14 +303,52 @@ mo_rank <- function(x, ...) {
 #' @export
 mo_taxonomy <- function(x, language = get_locale(),  ...) {
   x <- AMR::as.mo(x, ...)
-  base::list(kingdom = mo_kingdom(x, language = language),
-             phylum = mo_phylum(x, language = language),
-             class = mo_class(x, language = language),
-             order = mo_order(x, language = language),
-             family = mo_family(x, language = language),
-             genus = mo_genus(x, language = language),
-             species = mo_species(x, language = language),
-             subspecies = mo_subspecies(x, language = language))
+  base::list(kingdom = AMR::mo_kingdom(x, language = language),
+             phylum = AMR::mo_phylum(x, language = language),
+             class = AMR::mo_class(x, language = language),
+             order = AMR::mo_order(x, language = language),
+             family = AMR::mo_family(x, language = language),
+             genus = AMR::mo_genus(x, language = language),
+             species = AMR::mo_species(x, language = language),
+             subspecies = AMR::mo_subspecies(x, language = language))
+}
+
+#' @rdname mo_property
+#' @export
+mo_synonyms <- function(x, ...) {
+  x <- as.mo(x, ...)
+  IDs <- AMR::mo_property(x = x, property = "col_id", language = NULL)
+  syns <- lapply(IDs, function(col_id) {
+    res <- sort(AMR::microorganisms.old[which(AMR::microorganisms.old$col_id_new == col_id), "fullname"])
+    if (length(res) == 0) {
+      NULL
+    } else {
+      res
+    }
+  })
+  if (length(syns) > 1) {
+    names(syns) <- mo_fullname(x)
+    syns
+  } else {
+    unlist(syns)
+  }
+}
+
+#' @rdname mo_property
+#' @export
+mo_info <- function(x, language = get_locale(),  ...) {
+  x <- AMR::as.mo(x, ...)
+  info <- lapply(x, function(y)
+    c(mo_taxonomy(y, language = language),
+      list(synonyms = mo_synonyms(y),
+           url = unname(mo_url(y, open = FALSE)),
+           ref = mo_ref(y))))
+  if (length(info) > 1) {
+    names(info) <- mo_fullname(x)
+    info
+  } else {
+    info[[1L]]
+  }
 }
 
 #' @rdname mo_property
@@ -326,7 +367,7 @@ mo_url <- function(x, open = FALSE, ...) {
                              NA_character_))
 
   u <- df$url
-  names(u) <- mo_fullname(mo)
+  names(u) <- AMR::mo_fullname(mo)
   if (open == TRUE) {
     if (length(u) > 1) {
       warning("only the first URL will be opened, as `browseURL()` only suports one string.")
@@ -348,7 +389,7 @@ mo_property <- function(x, property = 'fullname', language = get_locale(), ...) 
     stop("invalid property: '", property, "' - use a column name of the `microorganisms` data set")
   }
 
-  t(mo_validate(x = x, property = property, ...), language = language)
+  translate_AMR(mo_validate(x = x, property = property, ...), language = language, only_unknown = TRUE)
 }
 
 mo_validate <- function(x, property, ...) {
@@ -376,12 +417,15 @@ mo_validate <- function(x, property, ...) {
   if (!all(x %in% pull(AMR::microorganisms, property))
       | Becker %in% c(TRUE, "all")
       | Lancefield %in% c(TRUE, "all")) {
-    exec_as.mo(x, property = property, ...)
-  } else {
-    if (property == "mo") {
-      return(structure(x, class = "mo"))
-    } else {
-      return(x)
-    }
+    x <- exec_as.mo(x, property = property, ...)
   }
+
+  if (property == "mo") {
+    return(structure(x, class = "mo"))
+  } else if (property == "col_id") {
+    return(as.integer(x))
+  } else {
+    return(x)
+  }
+
 }