+++ inst/doc/mo_property.R
@@ -1,8 +1,0 @@
-## ----setup, include = FALSE, results = 'markup'--------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#"
-)
-# set to original language (English)
-Sys.setlocale(locale = "C")
-