+++ NEWS.md
@@ -1,3 +1,64 @@
+# AMR 0.7.1
+
+#### New
+* Function `rsi_df()` to transform a `data.frame` to a data set containing only the microbial interpretation (S, I, R), the antibiotic, the percentage of S/I/R and the number of available isolates. This is a convenient combination of the existing functions `count_df()` and `portion_df()` to immediately show resistance percentages and number of available isolates:
+  ```r
+  septic_patients %>%
+    select(AMX, CIP) %>%
+    rsi_df()
+  #      antibiotic  interpretation      value  isolates
+  # 1   Amoxicillin              SI  0.4442636       546
+  # 2   Amoxicillin               R  0.5557364       683
+  # 3 Ciprofloxacin              SI  0.8381831      1181
+  # 4 Ciprofloxacin               R  0.1618169       228
+  ```
+* Support for all scientifically published pathotypes of *E. coli* to date (that we could find). Supported are: 
+
+  * AIEC (Adherent-Invasive *E. coli*) 
+  * ATEC (Atypical Entero-pathogenic *E. coli*) 
+  * DAEC (Diffusely Adhering *E. coli*) 
+  * EAEC (Entero-Aggresive *E. coli*) 
+  * EHEC (Entero-Haemorrhagic *E. coli*) 
+  * EIEC (Entero-Invasive *E. coli*) 
+  * EPEC (Entero-Pathogenic *E. coli*) 
+  * ETEC (Entero-Toxigenic *E. coli*) 
+  * NMEC (Neonatal Meningitis‐causing *E. coli*) 
+  * STEC (Shiga-toxin producing *E. coli*) 
+  * UPEC (Uropathogenic *E. coli*)
+  
+  All these lead to the microbial ID of *E. coli*:
+  ```r
+  as.mo("UPEC")
+  # B_ESCHR_COL
+  mo_name("UPEC")
+  # "Escherichia coli"
+  mo_gramstain("EHEC")
+  # "Gram-negative"
+  ```
+* Function `mo_info()` as an analogy to `ab_info()`. The `mo_info()` prints a list with the full taxonomy, authors, and the URL to the online database of a microorganism
+* Function `mo_synonyms()` to get all previously accepted taxonomic names of a microorganism
+
+#### Changed
+* Column names of output `count_df()` and `portion_df()` are now lowercase
+* Fixed bug in translation of microorganism names
+* Fixed bug in determining taxonomic kingdoms
+* Algorithm improvements for `as.ab()` and `as.mo()` to understand even more severely misspelled input
+* Function `as.ab()` now allows spaces for coercing antibiotics names
+* Added `ggplot2` methods for automatically determining the scale type of classes `mo` and `ab`
+* Added names of object in the header in frequency tables, even when using pipes
+* Prevented `"bacteria"` from getting coerced by `as.ab()` because Bacterial is a brand name of trimethoprim (TMP)
+* Fixed a bug where setting an antibiotic would not work for `eucast_rules()` and `mdro()`
+* Fixed a EUCAST rule for Staphylococci, where amikacin resistance would not be inferred from tobramycin
+* Removed `latest_annual_release` from the `catalogue_of_life_version()` function
+* Removed antibiotic code `PVM1` from the `antibiotics` data set as this was a duplicate of `PME`
+* Fixed bug where not all old taxonomic names would be printed, when using a vector as input for `as.mo()`
+* Manually added *Trichomonas vaginalis* from the kingdom of Protozoa, which is missing from the Catalogue of Life
+* Small improvements to `plot()` and `barplot()` for MIC and RSI classes
+* Allow Catalogue of Life IDs to be coerced by `as.mo()`
+
+#### Other
+* Fixed a note thrown by CRAN tests
+
 # AMR 0.7.0
 
 #### New